Adds a parallel info tracker for multilevel meshes.      - Fixes a broken test.
