+++ proteus/LinearAlgebraTools.py
@@ -358,21 +358,22 @@ class ParInfo_petsc4py:
     here as static class values.  Then ParVec and ParMat can
     use these values to create parallel objects later.
     """
-    par_bs = None
-    par_n = None
-    par_n_lst = None
-    par_N = None
-    par_nghost = None
-    par_nghost_lst = None
-    petsc_subdomain2global_petsc = None
-    subdomain2global = None
-    proteus2petsc_subdomain = None
-    petsc2proteus_subdomain = None
-    nzval_proteus2petsc = None
-    dim = None
-    mixed = False
 
-    @classmethod
+    def __init__(self):
+        self.par_bs = None
+        self.par_n = None
+        self.par_n_lst = None
+        self.par_N = None
+        self.par_nghost = None
+        self.par_nghost_lst = None
+        self.petsc_subdomain2global_petsc = None
+        self.subdomain2global = None
+        self.proteus2petsc_subdomain = None
+        self.petsc2proteus_subdomain = None
+        self.nzval_proteus2petsc = None
+        self.dim = None
+        self.mixed = False
+
     def print_info(cls):
         import Comm
         comm = Comm.get()
@@ -448,8 +449,7 @@ class ParMat_petsc4py(p4pyPETSc.Mat):
             self.setBlockSize(self.blockSize)
             self.subdomain2global = subdomain2global #no need to include extra block dofs?
         else:
-#            self.setType('seqaij') # ARB - in the current configuration this will give asm trouble on a single core.
-            self.setType('mpiaij')
+            self.setType('aij')
             self.setSizes([[par_n*self.blockSize,par_N*self.blockSize],[par_nc*self.blockSize,par_Nc*self.blockSize]],bsize=1)
             if self.blockSize > 1: #have to build in block dofs
                 subdomain2globalTotal = numpy.zeros((self.blockSize*subdomain2global.shape[0],),'i')
@@ -932,8 +932,23 @@ class TwoPhase_PCDInv_shell(InvOperatorShell):
         tmp2 = x.copy()
         tmp3 = x.copy()
 
-        self.kspQp_visc.solve(x,y)
-        self.kspQp_dens.solve(x,tmp1)
+        self.kspQp_visc = ChebyshevSemiIteration(self.Qp_visc,
+                                                 x,
+                                                 y,
+                                                 10,
+                                                 0.5,
+                                                 2.0)
+        self.kspQp_dens = ChebyshevSemiIteration(self.Qp_dens,
+                                                 x,
+                                                 y,
+                                                 10,
+                                                 0.5,
+                                                 2.0)
+
+        self.kspQp_visc.apply()
+        self.kspQp_dens.apply()
+#        self.kspQp_visc.solve(x,y)
+#        self.kspQp_dens.solve(x,tmp1)
         
         self.Np_rho.mult(tmp1,tmp2)
 
@@ -1203,3 +1218,141 @@ if __name__ == '__main__':
 #         mgits +=1
 #         mgv.apply(w,jits,jits,0,f[1:n-1],u[1:n-1])
 #         rnorm = l2Norm(resList[0])
+
+class IterativeMethod:
+    """  Base class for iterative solvers.
+
+    Parameters
+    ----------
+    A : matrix
+        System coefficient matrix.
+
+    b : vector
+        Right hand side.
+
+    x : vector
+        Initial guess for solution.
+
+    save_iterations: bool
+        Flag that allows the user to save iterations.
+    """
+    def __init__(self, A, b, x, save_iterations = False):
+        self.A = A
+        self.b = b
+        self.x_k = x.copy()
+        self.x_km1 = x.copy()
+        self.x_km1.zeroEntries()
+
+        self.n = self.A.size[0]
+        self.save_iterations = save_iterations
+        if self.save_iterations:
+            self.iteration_results = [self.x_k.getArray().copy()]
+
+    def _calc_residual(self, x_k):
+        Ax = x_k.copy()
+        self.A.mult(x_k,Ax)
+        Ax.aypx(-1., self.b)
+        return Ax
+
+class ChebyshevSemiIteration(IterativeMethod):
+    """ Class for implementing the ChebyshevSemiIteration. 
+    
+    Notes
+    -----
+    The Chebyshev semi-iteration was developed in the 1960s
+    by Golub and Varga.  It is an iterative technique for
+    solving linear systems Ax = b with the property of preserving
+    linearity with respect to the Krylov solves (see Wathen,
+    Rees 2009 - Chebyshev semi-iteration in preconditioning
+    for problems including the mass matrix).  This makes the method 
+    particularly well suited for solving sub-problems that arise in 
+    more complicated block preconditioners such as the Schur 
+    complement, provided one has an aprior bound on the systems 
+    eigenvalues (see Wathen 1987 - Realisitc eigenvalue bounds for 
+    Galerkin mass matrix).
+
+    When implementing this method it is important you first
+    have tight aprior bounds on the eigenvalues (denoted here as 
+    alpha and beta). This can be a challenge but, the references 
+    above do provide these results for many relevant mass matrices.
+    
+    Also, when implementing this method, the residual b - Ax0
+    will be preconditioned with the inverse of diag(A).  Your eigenvalue 
+    bounds should reflect the spectrum of this preconditioned system.
+
+    Arugments
+    ---------
+    A : petsc4py matrix
+        The linear system matrix
+
+    b : petsc4py vector
+        The righthand side vector
+
+    x : petsc4py vector
+        An initial guess for the solution
+
+    k : int
+        The desired number of iterations
+
+    alpha : float
+        A's smallest eigenvalue
+
+    beta : float
+        A's largest eigenvalue
+
+    save_iterations : bool
+        A flag indicating whether to store each solution iteration
+    """
+
+    def __init__(self, A, b, x, k, alpha, beta, save_iterations = False):
+        IterativeMethod.__init__(self, A, b, x, save_iterations)
+        self.k = k
+        self.alpha = alpha
+        self.beta = beta
+        self.relax_parameter = (self.alpha + self.beta) / 2.
+        self.rho = (self.beta - self.alpha) / (self.alpha + self.beta)
+        self.diag = A.getDiagonal().copy()
+        self.diag.scale(self.relax_parameter)
+        self.z = A.getDiagonal().copy()
+        
+    def apply(self):
+        for i in range(self.k):
+            w = 1./(1-(self.rho**2)/4.)
+            r = self._calc_residual(self.x_k)
+            # x_kp1 = w*(z + x_k - x_km1) + x_km1
+            self.z.pointwiseDivide(r, self.diag)
+            self.z.axpy(1., self.x_k)
+            self.z.axpy(-1., self.x_km1)
+            self.z.scale(w)
+            self.z.axpy(1., self.x_km1)
+            self.x_km1 = self.x_k.copy()
+            self.x_k = self.z.copy()
+            if self.save_iterations:
+                self.iteration_results.append(self.x_k.getArray().copy())
+        
+# The implementation that is commented out here was adopted from
+# the 1996 text Iterative Solution Methods by Owe Axelsson starting
+# on page 179.  As currently written, this algorithm requires
+# inputing the inverse diagonally preconditioned matrix A and b.  I'm
+# not sure this is the best approach, but I'd like to leave this code
+# in place for now in case it is useful in the future.
+
+    # def _calc_theta_ell(self, ell):
+    #     return ((2*ell + 1) / (2. * self.k) ) * math.pi
+
+    # def _calc_tau(self, ell):
+    #     theta = self._calc_theta_ell(ell)
+    #     one_over_tau = ( (self.beta - self.alpha) / 2. * math.cos(theta) +
+    #                      (self.beta + self.alpha) / 2.)
+    #     return 1. / one_over_tau
+
+    # def apply(self):
+    #     for i in range(self.k):
+    #         if i==0 and self.save_iterations:
+    #             self.iteration_results.append(self.x_k.getArray().reshape(self.n,1).copy())
+    #         elif i > 0:
+    #             tau = self._calc_tau(i-1)
+    #             resid = self._calc_residual(self.x_k)
+    #             self.x_k.axpy(-tau, resid)
+    #             if self.save_iterations:
+    #                 self.iteration_results.append(self.x_k.getArray().reshape(self.n,1).copy())
