+++ proteus/LinearSolvers.py
@@ -790,33 +790,34 @@ class SchurOperatorConstructor:
         import Comm
         comm = Comm.get()
         self.opBuilder.attachTPAdvectionOperator()
+        par_info = self.linear_smoother.L.pde.par_info
         if comm.size() == 1:
             self.two_phase_Cp_rho = ParMat_petsc4py(self.opBuilder.TPScaledAdvectionOperator,
-                                                    LAT.ParInfo_petsc4py.par_bs,
-                                                    LAT.ParInfo_petsc4py.par_n,
-                                                    LAT.ParInfo_petsc4py.par_N,
-                                                    LAT.ParInfo_petsc4py.par_nghost,
-                                                    LAT.ParInfo_petsc4py.subdomain2global)
+                                                    par_info.par_bs,
+                                                    par_info.par_n,
+                                                    par_info.par_N,
+                                                    par_info.par_nghost,
+                                                    par_info.subdomain2global)
         else:
             mixed = False
             if mixed == True:
                 self.petsc_two_phase_Cp_rho = self._initializeMat(self.opBuilder.TPScaledAdvectionOperator)
                 self.two_phase_Cp_rho = ParMat_petsc4py(self.petsc_two_phase_Cp_rho,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.petsc_subdomain2global_petsc,
                                                         pde=self.L.pde,
                                                         proteus_jacobian=self.opBuilder.TPScaledAdvectionOperator,
-                                                        nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                                        nzval_proteus2petsc=par_info.nzval_proteus2petsc)
             else:
                 self.two_phase_Cp_rho = ParMat_petsc4py(self.opBuilder.TPScaledAdvectionOperator,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.subdomain2global,
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.subdomain2global,
                                                         pde=self.L.pde)
         self.two_phase_Cp_rho_csr_rep = self.two_phase_Cp_rho.csr_rep
         self.two_phase_Cp_rho_csr_rep_local = self.two_phase_Cp_rho.csr_rep_local
@@ -832,33 +833,34 @@ class SchurOperatorConstructor:
         import Comm
         comm = Comm.get()
         self.opBuilder.attachLaplaceOperator()
+        par_info = self.linear_smoother.L.pde.par_info
         if comm.size() == 1:
             self.two_phase_Ap_inv = ParMat_petsc4py(self.opBuilder.TPInvScaledLaplaceOperator,
-                                                    LAT.ParInfo_petsc4py.par_bs,
-                                                    LAT.ParInfo_petsc4py.par_n,
-                                                    LAT.ParInfo_petsc4py.par_N,
-                                                    LAT.ParInfo_petsc4py.par_nghost,
-                                                    LAT.ParInfo_petsc4py.subdomain2global)
+                                                    par_info.par_bs,
+                                                    par_info.par_n,
+                                                    par_info.par_N,
+                                                    par_info.par_nghost,
+                                                    par_info.subdomain2global)
         else:
             mixed = False
             if mixed == True:
                 self.petsc_two_phase_Ap_inv = self._initializeMat(self.opBuilder.TPInvScaledLaplaceOperator)
                 self.two_phase_Ap_inv = ParMat_petsc4py(self.petsc_two_phase_Ap_inv,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.petsc_subdomain2global_petsc,
                                                         pde=self.L.pde,
                                                         proteus_jacobian=self.opBuilder.TPInvScaledLaplaceOperator,
-                                                        nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                                        nzval_proteus2petsc=par_info.nzval_proteus2petsc)
             else:
                 self.two_phase_Ap_inv = ParMat_petsc4py(self.opBuilder.TPInvScaledLaplaceOperator,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.subdomain2global,
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.subdomain2global,
                                                         pde=self.L.pde)                
         self.two_phase_Ap_inv_csr_rep = self.two_phase_Ap_inv.csr_rep
         self.two_phase_Ap_inv_csr_rep_local = self.two_phase_Ap_inv.csr_rep_local
@@ -874,33 +876,34 @@ class SchurOperatorConstructor:
         import Comm
         comm = Comm.get()
         self.opBuilder.attachScaledMassOperator()
+        par_info = self.linear_smoother.L.pde.par_info
         if comm.size() == 1:
             self.two_phase_Qp_scaled = ParMat_petsc4py(self.opBuilder.TPScaledMassOperator,
-                                                       LAT.ParInfo_petsc4py.par_bs,
-                                                       LAT.ParInfo_petsc4py.par_n,
-                                                       LAT.ParInfo_petsc4py.par_N,
-                                                       LAT.ParInfo_petsc4py.par_nghost,
-                                                       LAT.ParInfo_petsc4py.subdomain2global)
+                                                       par_info.par_bs,
+                                                       par_info.par_n,
+                                                       par_info.par_N,
+                                                       par_info.par_nghost,
+                                                       par_info.subdomain2global)
         else:
             mixed = False
             if mixed == True:
                 self.petsc_two_phase_Qp_scaled = self._initializeMat(self.opBuilder.TPScaledMassOperator)
                 self.two_phase_Qp_scaled = ParMat_petsc4py(self.petsc_two_phase_Qp_scaled,
-                                                           LAT.ParInfo_petsc4py.par_bs,
-                                                           LAT.ParInfo_petsc4py.par_n,
-                                                           LAT.ParInfo_petsc4py.par_N,
-                                                           LAT.ParInfo_petsc4py.par_nghost,
-                                                           LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                                           par_info.par_bs,
+                                                           par_info.par_n,
+                                                           par_info.par_N,
+                                                           par_info.par_nghost,
+                                                           par_info.petsc_subdomain2global_petsc,
                                                            pde=self.L.pde,
                                                            proteus_jacobian=self.opBuilder.TPScaledMassOperator,
-                                                           nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                                           nzval_proteus2petsc=par_info.nzval_proteus2petsc)
             else:
                 self.two_phase_Qp_scaled = ParMat_petsc4py(self.opBuilder.TPScaledMassOperator,
-                                                           LAT.ParInfo_petsc4py.par_bs,
-                                                           LAT.ParInfo_petsc4py.par_n,
-                                                           LAT.ParInfo_petsc4py.par_N,
-                                                           LAT.ParInfo_petsc4py.par_nghost,
-                                                           LAT.ParInfo_petsc4py.subdomain2global)
+                                                           par_info.par_bs,
+                                                           par_info.par_n,
+                                                           par_info.par_N,
+                                                           par_info.par_nghost,
+                                                           par_info.subdomain2global)
         self.two_phase_Qp_scaled_csr_rep = self.two_phase_Qp_scaled.csr_rep
         self.two_phase_Qp_scaled_csr_rep_local = self.two_phase_Qp_scaled.csr_rep_local
         return self.two_phase_Qp_scaled
@@ -915,33 +918,34 @@ class SchurOperatorConstructor:
         import Comm
         comm = Comm.get()
         self.opBuilder.attachInvScaledMassOperator()
+        par_info = self.linear_smoother.L.pde.par_info
         if comm.size() == 1:
             self.two_phase_Qp_inv = ParMat_petsc4py(self.opBuilder.TPInvScaledMassOperator,
-                                                    LAT.ParInfo_petsc4py.par_bs,
-                                                    LAT.ParInfo_petsc4py.par_n,
-                                                    LAT.ParInfo_petsc4py.par_N,
-                                                    LAT.ParInfo_petsc4py.par_nghost,
-                                                    LAT.ParInfo_petsc4py.subdomain2global)
+                                                    par_info.par_bs,
+                                                    par_info.par_n,
+                                                    par_info.par_N,
+                                                    par_info.par_nghost,
+                                                    par_info.subdomain2global)
         else:
             mixed = False
             if mixed == True:
                 self.petsc_two_phase_Qp_inv = self._initializeMat(self.opBuilder.TPInvScaledMassOperator)
                 self.two_phase_Qp_inv = ParMat_petsc4py(self.petsc_two_phase_Qp_inv,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.petsc_subdomain2global_petsc,
                                                         pde=self.L.pde,
                                                         proteus_jacobian=self.opBuilder.TPInvScaledMassOperator,
-                                                        nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                                        nzval_proteus2petsc=par_info.nzval_proteus2petsc)
             else:
                 self.two_phase_Qp_inv = ParMat_petsc4py(self.opBuilder.TPInvScaledMassOperator,
-                                                        LAT.ParInfo_petsc4py.par_bs,
-                                                        LAT.ParInfo_petsc4py.par_n,
-                                                        LAT.ParInfo_petsc4py.par_N,
-                                                        LAT.ParInfo_petsc4py.par_nghost,
-                                                        LAT.ParInfo_petsc4py.subdomain2global)                
+                                                        par_info.par_bs,
+                                                        par_info.par_n,
+                                                        par_info.par_N,
+                                                        par_info.par_nghost,
+                                                        par_info.subdomain2global)                
         self.two_phase_Qp_inv_csr_rep = self.two_phase_Qp_inv.csr_rep
         self.two_phase_Qp_inv_csr_rep_local = self.two_phase_Qp_inv.csr_rep_local
         return self.two_phase_Qp_inv
@@ -958,44 +962,45 @@ class SchurOperatorConstructor:
         import Comm
         comm = Comm.get()
         self.opBuilder.attachMassOperator()
+        par_info = self.linear_smoother.L.pde.par_info
         if comm.size() == 1:
-            if LAT.ParInfo_petsc4py.mixed is False:
+            if par_info.mixed is False:
                 self.Q = ParMat_petsc4py(self.opBuilder.MassOperator,
-                                         LAT.ParInfo_petsc4py.par_bs,
-                                         LAT.ParInfo_petsc4py.par_n,
-                                         LAT.ParInfo_petsc4py.par_N,
-                                         LAT.ParInfo_petsc4py.par_nghost,
-                                         LAT.ParInfo_petsc4py.subdomain2global)
+                                         par_info.par_bs,
+                                         par_info.par_n,
+                                         par_info.par_N,
+                                         par_info.par_nghost,
+                                         par_info.subdomain2global)
             else:
                 self.Q = ParMat_petsc4py(self.opBuilder.MassOperator,
                                          1,
-                                         LAT.ParInfo_petsc4py.par_n,
-                                         LAT.ParInfo_petsc4py.par_N,
-                                         LAT.ParInfo_petsc4py.par_nghost,
-                                         LAT.ParInfo_petsc4py.subdomain2global,
+                                         par_info.par_n,
+                                         par_info.par_N,
+                                         par_info.par_nghost,
+                                         par_info.subdomain2global,
                                          pde=self.L.pde)
         else:
-            if LAT.ParInfo_petsc4py.mixed is True:
+            if par_info.mixed is True:
                 self.petsc_Q = self._initializeMat(self.opBuilder.MassOperator)
                 self.Q = ParMat_petsc4py(self.opBuilder.MassOperator,
                                          1,
-                                         LAT.ParInfo_petsc4py.par_n,
-                                         LAT.ParInfo_petsc4py.par_N,
-                                         LAT.ParInfo_petsc4py.par_nghost,
-                                         LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                         par_info.par_n,
+                                         par_info.par_N,
+                                         par_info.par_nghost,
+                                         par_info.petsc_subdomain2global_petsc,
                                          pde=self.L.pde,
                                          proteus_jacobian=self.opBuilder.MassOperator,
-                                         nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                         nzval_proteus2petsc=par_info.nzval_proteus2petsc)
             else:
                 self.Q = ParMat_petsc4py(self.petsc_Q,
-                                         LAT.ParInfo_petsc4py.par_bs,
-                                         LAT.ParInfo_petsc4py.par_n,
-                                         LAT.ParInfo_petsc4py.par_N,
-                                         LAT.ParInfo_petsc4py.par_nghost,
-                                         LAT.ParInfo_petsc4py.petsc_subdomain2global_petsc,
+                                         par_info.par_bs,
+                                         par_info.par_n,
+                                         par_info.par_N,
+                                         par_info.par_nghost,
+                                         par_info.petsc_subdomain2global_petsc,
                                          pde=self.L.pde,
                                          proteus_jacobian=self.opBuilder.MassOperator,
-                                         nzval_proteus2petsc=LAT.ParInfo_petsc4py.nzval_proteus2petsc)
+                                         nzval_proteus2petsc=par_info.nzval_proteus2petsc)
         self.Q_csr_rep = self.Q.csr_rep
         self.Q_csr_rep_local = self.Q.csr_rep_local
         return self.Q
@@ -1066,8 +1071,6 @@ class SchurOperatorConstructor:
         ----------
         output_matrix : bool 
             Determines whether matrix should be exported.
-        recalculate : bool
-            Flag indicating whther matrix should be rebuilt every time it's used
 
         Returns
         -------
@@ -3273,140 +3276,3 @@ class OperatorConstructor_oneLevel(OperatorConstructor):
                                                                           A)
 
 
-class IterativeMethod:
-    """  Base class for iterative solvers.
-
-    Parameters
-    ----------
-    A : matrix
-        System coefficient matrix.
-
-    b : vector
-        Right hand side.
-
-    x : vector
-        Initial guess for solution.
-
-    save_iterations: bool
-        Flag that allows the user to save iterations.
-    """
-    def __init__(self, A, b, x, save_iterations = False):
-        self.A = A
-        self.b = b
-        self.x_k = x.copy()
-        self.x_km1 = x.copy()
-        self.x_km1.zeroEntries()
-
-        self.n = self.A.size[0]
-        self.save_iterations = save_iterations
-        if self.save_iterations:
-            self.iteration_results = [self.x_k.getArray().copy()]
-
-    def _calc_residual(self, x_k):
-        Ax = x_k.copy()
-        self.A.mult(x_k,Ax)
-        Ax.aypx(-1., self.b)
-        return Ax
-
-class ChebyshevSemiIteration(IterativeMethod):
-    """ Class for implementing the ChebyshevSemiIteration. 
-    
-    Notes
-    -----
-    The Chebyshev semi-iteration was developed in the 1960s
-    by Golub and Varga.  It is an iterative technique for
-    solving linear systems Ax = b with the property of preserving
-    linearity with respect to the Krylov solves (see Wathen,
-    Rees 2009 - Chebyshev semi-iteration in preconditioning
-    for problems including the mass matrix).  This makes the method 
-    particularly well suited for solving sub-problems that arise in 
-    more complicated block preconditioners such as the Schur 
-    complement, provided one has an aprior bound on the systems 
-    eigenvalues (see Wathen 1987 - Realisitc eigenvalue bounds for 
-    Galerkin mass matrix).
-
-    When implementing this method it is important you first
-    have tight aprior bounds on the eigenvalues (denoted here as 
-    alpha and beta). This can be a challenge but, the references 
-    above do provide these results for many relevant mass matrices.
-    
-    Also, when implementing this method, the residual b - Ax0
-    will be preconditioned with the inverse of diag(A).  Your eigenvalue 
-    bounds should reflect the spectrum of this preconditioned system.
-
-    Arugments
-    ---------
-    A : petsc4py matrix
-        The linear system matrix
-
-    b : petsc4py vector
-        The righthand side vector
-
-    x : petsc4py vector
-        An initial guess for the solution
-
-    k : int
-        The desired number of iterations
-
-    alpha : float
-        A's smallest eigenvalue
-
-    beta : float
-        A's largest eigenvalue
-
-    save_iterations : bool
-        A flag indicating whether to store each solution iteration
-    """
-
-    def __init__(self, A, b, x, k, alpha, beta, save_iterations = False):
-        IterativeMethod.__init__(self, A, b, x, save_iterations)
-        self.k = k
-        self.alpha = alpha
-        self.beta = beta
-        self.relax_parameter = (self.alpha + self.beta) / 2.
-        self.rho = (self.beta - self.alpha) / (self.alpha + self.beta)
-        self.diag = A.getDiagonal().copy()
-        self.diag.scale(self.relax_parameter)
-        self.z = A.getDiagonal().copy()
-        
-    def apply(self):
-        for i in range(self.k):
-            w = 1./(1-(self.rho**2)/4.)
-            r = self._calc_residual(self.x_k)
-            # x_kp1 = w*(z + x_k - x_km1) + x_km1
-            self.z.pointwiseDivide(r, self.diag)
-            self.z.axpy(1., self.x_k)
-            self.z.axpy(-1., self.x_km1)
-            self.z.scale(w)
-            self.z.axpy(1., self.x_km1)
-            self.x_km1 = self.x_k.copy()
-            self.x_k = self.z.copy()
-            if self.save_iterations:
-                self.iteration_results.append(self.x_k.getArray().copy())
-        
-# The implementation that is commented out here was adopted from
-# the 1996 text Iterative Solution Methods by Owe Axelsson starting
-# on page 179.  As currently written, this algorithm requires
-# inputing the inverse diagonally preconditioned matrix A and b.  I'm
-# not sure this is the best approach, but I'd like to leave this code
-# in place for now in case it is useful in the future.
-
-    # def _calc_theta_ell(self, ell):
-    #     return ((2*ell + 1) / (2. * self.k) ) * math.pi
-
-    # def _calc_tau(self, ell):
-    #     theta = self._calc_theta_ell(ell)
-    #     one_over_tau = ( (self.beta - self.alpha) / 2. * math.cos(theta) +
-    #                      (self.beta + self.alpha) / 2.)
-    #     return 1. / one_over_tau
-
-    # def apply(self):
-    #     for i in range(self.k):
-    #         if i==0 and self.save_iterations:
-    #             self.iteration_results.append(self.x_k.getArray().reshape(self.n,1).copy())
-    #         elif i > 0:
-    #             tau = self._calc_tau(i-1)
-    #             resid = self._calc_residual(self.x_k)
-    #             self.x_k.axpy(-tau, resid)
-    #             if self.save_iterations:
-    #                 self.iteration_results.append(self.x_k.getArray().reshape(self.n,1).copy())
