+++ setup.py
@@ -463,7 +463,6 @@ setup(name='proteus',
                     'proteus/tests/solver_tests_slow/comparison_files/drivenCavityStokes_expected.h5']),
                   (os.path.join(proteus_install_path,'tests','matrix_constructor','comparison_files'),
                    ['proteus/tests/matrix_constructor/comparison_files/mass_reference_c0p1_2D.txt',
-                    'proteus/tests/matrix_constructor/comparison_files/mass_reference_c0p1.txt',
                     'proteus/tests/matrix_constructor/comparison_files/mass_reference_TH_2D.npy']),
                   (os.path.join(proteus_install_path,'tests','post_processing','import_modules'),
                    ['proteus/tests/post_processing/import_modules/reference_simplex_keep.ele',
